ML Model/requirements.txt
Flask
Flask-Cors
azure-cognitiveservices-vision-computervision